feat:  introduce one-to-one on member and moderator &  reintroduce admin model (which meant removing roles table)
I need a way to ensure a user can be a member and a moderator at the same.
having the polymorphic relationship on Moderator means a user could either be a Member or Moderator, but not both.
For that I need a one-to-one relationship.
Having the roles table is redundant since I have Moderator and Member models